+++ docs/roadmap/roadmap.md
@@ -6,11 +6,14 @@
 - [x] vfs watches (inject/update on change)
 - [x] prompt management (ala claw)
 - [x] search & web requests ([P66 first cut](p66-minimal-web-access.md))
-- direct remote MCP ([P67 proposed](p67-direct-remote-mcp.md))
-- remote MCP registry/linking ([P68 proposed](p68-remote-mcp-registry-linking.md))
-- generic auth/token broker ([P69 proposed](p69-generic-auth-token-broker.md))
-- external integration surface: api schema, ts/py clients, idempotent runs, long-poll events ([P70 proposed](p70-external-integration-surface.md))
-- messaging channel gateway: activation policy, room context, delivery outbox, messaging tools, media input ([P71 G1–G5 first cuts done](p71-messaging-channel-gateway.md))
+- [x] direct remote MCP ([P67 proposed](p67-direct-remote-mcp.md))
+- [x] remote MCP registry/linking ([P68 proposed](p68-remote-mcp-registry-linking.md))
+- [x] generic auth/token broker ([P69 proposed](p69-generic-auth-token-broker.md))
+- [x] external integration surface: api schema, ts/py clients, idempotent runs, long-poll events ([P70 proposed](p70-external-integration-surface.md))
+- [x] messaging channel gateway: activation policy, room context, delivery outbox, messaging tools, media input ([P71 G1–G5 first cuts done](p71-messaging-channel-gateway.md))
+    - authetication/password/allowlists
+    - session binding & consistency (what about /new)
+    - cross session knowledge
 - (go live)
 - timers/schedules/cron
 - vms/sandboxes
